requirements.txt
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
curl_cffi==0.13.0
Flask==3.1.2
frozendict==2.4.7
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
keyboard==0.13.5
MarkupSafe==3.0.3
multitasking==0.0.12
numpy==2.2.6
pandas==2.3.3
peewee==3.19.0
platformdirs==4.5.1
protobuf==6.33.5
psycopg2-binary==2.9.11
pycparser==3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
six==1.17.0
soupsieve==2.8.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
websockets==16.0
Werkzeug==3.1.5
yfinance==1.1.0
